Make the OCCreds part of the OcCredentials class
In testing we kept seing random crashs due to segfaults.  The segfaults
are due to the fact that the native code is expecting the freed resources
to still be there.

When the OcCredentials class is created the OCCreds is now passed in and becomes
a member variable. It is no longer instantly freed. To avoid memory leaks the
OCCreds.freeCreds() function was moved into the Finalizer of the OcCredentials
class. There is sometimes reliablility problems with Finalizers that require
by passing the GC to manualy free the native memory.  The OcCredentials.delete()
method can be used to manualy free the native memory.

Some OCCreds are owned by the iotivity-lite stack and should not be freed
a boolean input parameter memoryOwner. True means Java owns the memory and
it can be freed by the GC, false means the iotivity-lite stack owns the memory.

When possible  the `get` functions were updated to get the value requested directly
from OCCreds. This change worked its way into OcCredential, OcCredPrivateData, and
OcCredPublicData classes.

While updating OcCredPrivateData and OcCredPublicData I saw that the setDerData was
never called so the `byte[] derData` member verable was alway unassigned. When I
went to remove the `getDerData()` member function I saw that it was being called
in the TrustAnchorAdaptor and it it was not assigned the pemData was converted to
derData. The converson code found in TrustAnchorAdapter was reused in the
OcCredPrivateData and OcCredPublicData classes to do the conversion when requested.

Since getCredId method returns an `int` not an `Integer` the return type was changed
to `int`.  In the code where the Integer.toString() method was used the code was
updated to take an `int` instead.

Signed-off-by: George Nash <george.nash@intel.com>